Reuse AndroidSymbol on web and rebuild
