--- packages/expo-symbols/src/SymbolView.web.tsx
@@ -1,5 +1,0 @@
-import { SymbolViewProps } from './SymbolModule.types';
-
-export function SymbolView(props: SymbolViewProps) {
-  return props.fallback;
-}
